uncompressed: pulled ComponentInterleaveDecoder into its own file
