+++ libheif/CMakeLists.txt
@@ -193,7 +193,9 @@ if (WITH_UNCOMPRESSED_CODEC)
             codecs/uncompressed/unc_dec.h
             codecs/uncompressed/unc_dec.cc
             codecs/uncompressed/decoder_abstract.h
-            codecs/uncompressed/decoder_abstract.cc)
+            codecs/uncompressed/decoder_abstract.cc
+            codecs/uncompressed/decoder_component_interleave.h
+            codecs/uncompressed/decoder_component_interleave.cc)
 endif ()
 
 write_basic_package_version_file(${PROJECT_NAME}-config-version.cmake COMPATIBILITY ExactVersion)
